Move Notification class into CellEditor namespace
